requirements.txt
requests[security]
python-keystoneclient==2.1.2
python-novaclient==3.2.0
python-cinderclient==1.5.0
python-glanceclient==1.2.0
python-neutronclient==4.0.0
python-ceilometerclient==2.2.1

python-openstackclient
